Small change to the rules for better playability
Now, players may only choose a tile from the bottom row after they
reached the minimum score. This removes the issue of players not knowing
in advance if selecting a tile on the bottom row will actually be enough
to win. There is the preview feature, but that's only a bandaid. Now,
there is no ambiguity left. However, the game is also a bit more
difficult as a result.